requirements.txt
numpy==2.3.2
pandas==2.3.1
bottleneck==1.5.0
numexpr==2.11.0
ft-pandas-ta==0.3.15
ta-lib==0.5.5
technical==1.5.2
ccxt==4.4.99
cryptography==45.0.6
aiohttp==3.12.15
SQLAlchemy==2.0.42
python-telegram-bot==22.3
httpx>=0.24.1
humanize==4.12.3
cachetools==6.1.0
requests==2.32.4
urllib3==2.5.0
certifi==2025.8.3
jsonschema==4.25.0
tabulate==0.9.0
pycoingecko==3.2.0
jinja2==3.1.6
joblib==1.5.1
rich==14.1.0
pyarrow
